Split activity db.rs into domain-specific submodules
Extract 5 domain modules from the monolithic db.rs (3,925 lines → 615 lines):
- claims.rs: Issue claim registry operations and tests
- cost_analytics.rs: Budget configuration, cost queries, and runway projection
- metrics.rs: Task lifecycle, token usage, and productivity tracking
- prompts.rs: Prompt git changes and GitHub event correlation
- quality.rs: Quality metrics, PR review status, and rework tracking

ActivityDb retains core operations (input/output recording, schema init) and
delegates to submodules via thin wrapper methods. Public API unchanged.

Closes #1881
